WIP: Show "fix" for ContactCard/set update on group with members
column_text_to_buf() is supposed to avoid pointer re-use in sqlite3, but it
uses what looks like a tiny "ring buffer" of bufs attached to the open db.

These bufs get reused when processing a ContactCard with members...

It's unclear to me how this problem doesn't affect Contacts too...

This fix is not a real fix as it would leak memory, but maybe points to a
possible solution. I don't know!

_card_set_update has:

    +    /* make a working copy of cdata -
    +     * conversion of vCard to JSContact might do group member lookups
    +     * and thus overwrite our cdata
    +     */
    +    cdata = memcpy(&mycdata, cdata, sizeof(mycdata));

... which is maybe supposed to solve this, but that's a shallow copy, meaning
the cdata->vcard_uid pointer is still the base pointer which gets overwritten,
so maybe that's where the real fix should be? Unsure.