Cassandane::Cyrus::TestCase: Add JMAP::Tester(::WebSocket) support
Cyrus supports websockets but cass has no testing capabilities for it.

This adds that through the use of JMAP::Tester::WebSocket.

At the same time, this adds JMAP::Tester incase we decide to use it for
anything in the future, as it provides a perhaps easier interface to use
than Mail::JMAPTalk for some things.